requirements/base.txt
Django==4.2.11
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
django-autoslug==1.9.9
django-countries==7.5.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.32
qrcode[pil]==7.4.2
drf-yasg==1.21.7
pillow==10.2.0
argon2-cffi==23.1.0
pytz==2024.1
redis==5.0.3
celery==5.3.6
flower==2.0.1
django-celery-email==3.0.0
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
django-haystack==3.2.1
Whoosh==2.7.4
drf-haystack==1.8.13
whitenoise==6.6.0
django-allauth[socialaccount]==0.63.0
django-allauth-2fa==0.11.1
django-otp==1.5.0
pyotp==2.9.0
django-two-factor-auth==1.16.0
social-auth-app-django==5.4.0